+++ velesresearch/website_template/src/SurveyComponent.jsx
@@ -93,6 +93,12 @@ async function handleResults(survey, completedHtml) {
   }
 }
 
+// {% customFunctions %}
+
+// placeholder
+
+// {% end customFunctions %}
+
 registerCustomFunctions();
 
 function SurveyComponent() {
@@ -120,6 +126,12 @@ function SurveyComponent() {
   survey.onAfterRenderSurvey.add((sender, options) => {
     document.body.style.setProperty("--sjs-general-backcolor-dim", document.getElementsByClassName("sd-root-modern")[0].style.getPropertyValue("--sjs-general-backcolor-dim"));
   });
+
+  // {% customCode %}
+
+  // placeholder
+
+  // {% end customCode %}
 
   survey.onComplete.add(sender => handleResults(sender, completedHtml));
   return <Survey model={survey} />;