+++ NEWS.md
@@ -1,3 +1,4 @@
 # future velesresearch version
 
 * `yarn` is now called directly through `subprocess.run()`. This should resolve some problems in Windows. `pynpm` dependency has been removed.
+* Added `customCode` and `customFunctions` arguments for direct javascript use. Also added `getJS()` helper to get the code from a file.